requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
asttokens==3.0.1
attrs==26.1.0
auth0-python==5.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.4.2
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.6
chromadb==1.5.5
click==8.3.1
colorama==0.4.6
crewai==0.80.0
crewai-tools==0.76.0
cryptography==46.0.6
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
durationpy==0.10
executing==2.2.1
fastuuid==0.14.0
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.73.1
greenlet==3.3.2
groq==0.13.1
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.8.0
identify==2.6.18
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
instructor==1.12.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json_repair==0.58.7
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.1.1
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
lance-namespace==0.6.1
lance-namespace-urllib3-client==0.6.1
lancedb==0.30.0
langchain==0.2.16
langchain-community==0.2.12
langchain-core==0.2.43
langchain-groq==0.1.9
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
litellm==1.80.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
neo4j==5.27.0
networkx==3.6.1
nodeenv==1.10.0
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.24.4
openai==1.109.1
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
overrides==7.7.0
packaging==24.2
parso==0.8.6
platformdirs==4.9.4
pre_commit==4.5.1
primp==1.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
pure_eval==0.2.3
pyarrow==23.0.1
pybase64==1.4.3
pycparser==3.0
pydantic==2.10.4
pydantic-settings==2.13.1
pydantic_core==2.27.2
Pygments==2.19.2
PyJWT==2.12.1
pypdf==6.9.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-discovery==1.2.1
python-docx==1.2.0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2026.1.post1
pyvis==0.3.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.33.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
setuptools==69.5.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
stack-data==0.6.3
stagehand==3.18.0
sympy==1.14.0
tenacity==8.5.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.1
tomli_w==1.2.0
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uv==0.11.2
uvicorn==0.42.0
virtualenv==21.2.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
yarl==1.23.0
youtube-transcript-api==1.2.4
zipp==3.23.0
